+++ src/components/ui/gameViewer/icons.js
@@ -1,6 +1,7 @@
 import EnvService from '../../../services/api/envService';
 
-const path = `${EnvService.getRoot()}assets/contentsw/games/`;
+const path = `${EnvService.getRoot()}assets/games/`;
+const safePath = `${EnvService.getRoot()}assets/games2/`;
 
 export const GameLabelIcons = {
     wordgame: {
@@ -14,5 +15,23 @@ export const GameLabelIcons = {
     },
     barslot: {
         source: path+'barslot.png'
+    }
+};
+
+export const SafeGameIcons = {
+    destinationBall: {
+        source: safePath+'destinationball.jpg'
+    },
+    dual1: {
+        source: safePath+'dual1.jpg'
+    },
+    nottretank: {
+        source: safePath+'nottretank.jpg'
+    },
+    tictoevue: {
+        source: safePath+'tictoevue.jpg'
+    },
+    parkingcar: {
+        source: safePath+'parkingcar.jpg'
     }
 };