+++ src/pages/safe/games/safeGames.js
@@ -1,0 +1,49 @@
+import { SafeGameIcons } from '../../../components/ui/gameViewer/icons';
+
+export const gmSfGames = [
+    { 
+        id: 'destinationBall', 
+        title: 'Destination Ball', 
+        link: 'https://www.norwaydict.com/xdual1/',
+        params: '&player=1&mode=real',
+        icon: SafeGameIcons.destinationBall,
+        w: 550,
+        h: 530
+    },
+    {
+        id: 'dual1',
+        title: 'Dual1',
+        link: 'https://www.norwaydict.com/xdual1/',
+        params: '&player=1&mode=real',
+        icon: SafeGameIcons.dual1,
+        w: 600,
+        h: 500
+    },
+    {
+        id: 'nottretank',
+        title: 'Nottretank',
+        link: 'https://norwaydict.com/html5games/nolettank',
+        params: '&player=1&mode=real',
+        icon: SafeGameIcons.nottretank,
+        w: 600,
+        h: 500
+    },
+    {
+        id: 'tictoevue',
+        title: 'TictoeVue',
+        link: 'https://norwaydict.com/vue-tictac-toe/',
+        params: '&player=1&mode=real',
+        icon: SafeGameIcons.tictoevue,
+        w: 600,
+        h: 500
+    },
+    {
+        id: 'parkingcar',
+        title: 'Parkingcar',
+        link: 'https://norwaydict.com/parkingcar',
+        params: '&player=1&mode=real',
+        icon: SafeGameIcons.parkingcar,
+        w: 600,
+        h: 500
+    }
+];