+++ src/pages/safe/index.js
@@ -1,5 +1,6 @@
 import './safeKing/index';
 import './editor/editorPage';
 import './banner/banner';
+import './games/game';
 import './welcome';
 import './tabs';