+++ src/pages/safe/sets.js
@@ -3,6 +3,7 @@ export const SafePageTabs = {
         game: 'game-btn',
         welcome: 'welcome-btn',
         editor: 'text-editor-btn',
+        games: 'safe-games-btn',
         banner: 'banner-btn'
     }
 };